requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.11.2
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.104.1
git-filter-repo==2.47.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
lxml==6.0.2
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
playwright==1.58.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.24.0
xlrd==2.0.2